+++ time2vec/__init__.py
@@ -1,3 +1,3 @@
-from time2vec.layer import EfficientTime2Vec
+from time2vec.layer import Time2Vec
 
-__all__ = ["EfficientTime2Vec"]
+__all__ = ["Time2Vec"]